빌드에러 수정, 적 위 이름 안뜨게 수정

+++ Source/Project_T/Private/InGame/IG_GameState.cpp
@@ -174,7 +174,7 @@ void AIG_GameState::EndGame()
 
 void AIG_GameState::ExpandPool()
 {
-	int32 newSize{ objectPoolDefaultSize * 1.5 };
+	int32 newSize{ static_cast<int32>(objectPoolDefaultSize * 1.5) };
 	for (int32 i = objectPoolDefaultSize; i < newSize; ++i)
 	{
 		auto enemy{ SpawnEnemy() };
